requirements.txt
kafka-python==2.0.2
elasticsearch==7.17.0
PyYAML==6.0
requests==2.31.0
Flask==2.3.2
prometheus_client==0.16.0